Create file to generate graphs with a Body measurement file
Created with AI with the prompt:
```
Create a graph of the csv file in the data folder. Separate the graphs by the type. So if a type is weight, then there should be a graph for weight. Do this for all the different types. Align the timeframes across all graphs. Output the
graphs to a file in the out folder.
```